Simplify how resources are loaded.
Instead of using a viewlet which does a lot of logic to determine what text fields are available and whether they have field-specific configuration,
field-specific configuration is now served with the HTML in a data attribute on the textarea. This means that instead of building tiny_mce_gzip.js
for each context, we can now build it once on the portal root and cache it.

This has various benefits:
- Fixes some regressions that made the editor fail to initialize if it wasn't associated with a known field.
- Improved cachability
- Less code