Added tablerowstyles setting in control panel.
These styles are available when you are setting the table row
properties.  Until now, all styles were shown here, most of which are
not useful for table rows, and they were shown in an ugly and wrong
way.

Note that this went wrong probably since Products.TinyMCE 1.3.
In row.js this does:

    addClassesToList('class', 'table_row_styles');

But `table_row_styles` was not available in the json configuration,
so addClassesToList fell back to the `advanced_styles`,
which resulted in a dictionary being shown as option.